Implemented the getProperty method from the Proxy interface.  This was required to use these objects in Trees.
The implementation could be optimized more but I only have one small tree in my application so I am not going to invest the time currently to find the most optimal solution.

It is recommended not to access these lists using the [int] operator because it can be much slower than using an iterator or using a for...each loop.